+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.9.1]
+### Fixed
+* Fixed bug with distinguishing of widget and shortcode on the page
+
 ## [2.9]
 ### Tested
 * Tested with WooCommerce 4.5